stage1/pyproject.toml
[project]
name = "stage1-ingest-augment"
version = "0.1.0"
description = "Stage 1: Ingest and Augment pipeline stage"
requires-python = ">=3.10"
dependencies = [
    "pipeline-common @ file://${PROJECT_ROOT}/pipeline_common",
    "boto3>=1.34.0",
    "botocore>=1.34.0",
    "python-dotenv>=1.0.0",
    "pydantic>=2.5.0",
    "albumentations>=1.3.0",
    "opencv-python-headless>=4.8.0",
    "Pillow>=10.0.0",
    "label-studio-sdk>=1.0.0",
]
optional-dependencies = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "ruff>=0.1.0",
    "mypy>=1.7.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.10"

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
